requirements.txt
celery==4.1.0
dj-database-url==0.4.2
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
django-solo==1.1.2
gunicorn==19.7.1
psycopg2==2.7.3.2
redis==2.10.6
requests==2.18.4
whitenoise==3.3.1
xhtml2pdf==0.0.6
html5lib==1.0b8
python-slugify==1.2.4
django-allauth==0.34.0